// src/app/Models/paciente.ts
export class Paciente {
  id: String;
  idade: String;
  email: String;
  senha: String;
  cintura: String;
  aumentou: String;
  imc: Number;
  nome: String;
  genero: String;
  diadiag: String;
  mesdiag: String;
  anodiag: String;
  pesoatual: String;
  pesohabitual: String;
  estatura: String;
  data: Date;
}